Implement batch remote streaming joining.
If a data set must be joined with another data set, and they are not both rasters or compatible SQL-based data sets, the join is now performed as follows: for each batch of rows (default stream batch size), a filter() query is performed on the foreign data set to retrieve all rows that can possibly match with the rows from the source. Then the join is performed in-memory on this subset, and the result is streamed.

The good thing about this is that it uses indexing on the foreign side (which is great if it's a database), and doesn't require downloading the full foreign data set to memory. Expressions are optimized to some extent (i.e. A=1 OR A=2 OR A=3 is written as A IN (1,2,3)) so that the database should be able to filter efficiently.

This method is memory-efficient for joins that have a 1:1 relationship between left and right, or joins where rows from the left side all map to the same row on the right side (m:n where m>n). It breaks down for joins where a single row on the left side maps to a high number of rows on the right side (m:n where n>>m). However, there is no good alternative for such joins apart from performing it in-database (which will be tried before QBEJoinTransformer is put to work).